+++ packages/lumx-react/src/components/comment-block/CommentBlock.tsx
@@ -1,4 +1,4 @@
-import React, { Children, forwardRef, KeyboardEvent, KeyboardEventHandler, ReactNode } from 'react';
+import React, { Children, forwardRef, ReactNode } from 'react';
 
 import classNames from 'classnames';
 
@@ -6,7 +6,6 @@ import { Avatar, Size, Theme, Tooltip } from '@lumx/react';
 import { Comp, GenericProps, HasTheme, ValueOf } from '@lumx/react/utils/type';
 import { getRootClassName, handleBasicClasses } from '@lumx/react/utils/className';
 
-import isFunction from 'lodash/isFunction';
 import { AvatarProps } from '../avatar/Avatar';
 
 /**
@@ -28,9 +27,9 @@ export interface CommentBlockProps extends GenericProps, HasTheme {
     avatarProps: AvatarProps;
     /** Comment block replies. */
     children?: ReactNode;
-    /** Comment date with the minimal timestamp informations (xx minutes, x hours, yesterday, 6 days, Month Day, Month Day Year)*/
-    date: string;
-    /** Comment date with the full timestamp informations (day, month, year, time) */
+    /** Comment date with the minimal timestamp information (xx minutes, x hours, yesterday, 6 days, Month Day, Month Day Year)*/
+    date?: string;
+    /** Comment date with the full timestamp information (day, month, year, time) */
     fullDate?: string;
     /** Whether the component has actions to display or not. */
     hasActions?: boolean;
@@ -41,12 +40,21 @@ export interface CommentBlockProps extends GenericProps, HasTheme {
     /** Whether the comment is relevant or not. */
     isRelevant?: boolean;
     /** Comment author name. */
-    name: string;
-    /** On click callback. */
+    name?: React.ReactNode;
+    /**
+     * On click callback.
+     * @deprecated Use avatarProps instead and/or inject a clickable component in `name`
+     */
     onClick?(): void;
-    /** On mouse enter callback. */
+    /**
+     * On mouse enter callback.
+     * @deprecated Use avatarProps instead and/or inject a clickable component in `name`
+     */
     onMouseEnter?(): void;
-    /** On mouse leave callback. */
+    /**
+     * On mouse leave callback.
+     * @deprecated Use avatarProps instead and/or inject a clickable component in `name`
+     */
     onMouseLeave?(): void;
     /** Comment content. */
     text: ReactNode | string;
@@ -100,11 +108,6 @@ export const CommentBlock: Comp<CommentBlockProps, HTMLDivElement> = forwardRef(
         variant,
         ...forwardedProps
     } = props;
-    const enterKeyPress: KeyboardEventHandler<HTMLElement> = (evt: KeyboardEvent<HTMLElement>) => {
-        if (evt.key === 'Enter' && isFunction(onClick)) {
-            onClick();
-        }
-    };
     const hasChildren = Children.count(children) > 0;
 
     return (
@@ -125,29 +128,23 @@ export const CommentBlock: Comp<CommentBlockProps, HTMLDivElement> = forwardRef(
         >
             <div className={`${CLASSNAME}__wrapper`}>
                 <div className={`${CLASSNAME}__avatar`}>
-                    <Avatar
-                        {...avatarProps}
-                        size={Size.m}
-                        tabIndex={onClick ? 0 : -1}
-                        onClick={onClick}
-                        onKeyPress={enterKeyPress}
-                    />
+                    <Avatar {...avatarProps} size={Size.m} onClick={onClick} />
                 </div>
 
                 <div className={`${CLASSNAME}__container`}>
                     <div className={`${CLASSNAME}__content`}>
                         <div className={`${CLASSNAME}__meta`}>
-                            <span
-                                className={`${CLASSNAME}__name`}
-                                onClick={onClick}
-                                onKeyPress={enterKeyPress}
-                                onMouseEnter={onMouseEnter}
-                                onMouseLeave={onMouseLeave}
-                                role="button"
-                                tabIndex={onClick ? 0 : -1}
-                            >
-                                {name}
-                            </span>
+                            {name && (
+                                // eslint-disable-next-line jsx-a11y/click-events-have-key-events,jsx-a11y/no-static-element-interactions
+                                <span
+                                    className={`${CLASSNAME}__name`}
+                                    onClick={onClick}
+                                    onMouseEnter={onMouseEnter}
+                                    onMouseLeave={onMouseLeave}
+                                >
+                                    {name}
+                                </span>
+                            )}
                             {headerActions && <span className={`${CLASSNAME}__header-actions`}>{headerActions}</span>}
                         </div>